requirements.txt
google-genai>=0.2.0
fastmcp>=0.1.0
pipedream>=0.3.0
python-dotenv>=1.0.0
websockets>=12.0
python-multipart>=0.0.6

